Move query preprocessing to its own mixin class.
